refactor: remove parallel routes
