--- src/app/dashboard/events/[id]/@modals/[...catchAll]/page.tsx
@@ -1,3 +1,0 @@
-export default function CatchAll() {
-  return null;
-}
