--- src/app/dashboard/events/[id]/@modals/page.tsx
@@ -1,3 +1,0 @@
-export default function Default() {
-  return null;
-}
