+++ src/app/dashboard/events/[id]/forms/(steps)/general-info.tsx
