+++ src/app/dashboard/events/[id]/forms/page.tsx
@@ -1,29 +1,24 @@
-import { SquarePlus } from "lucide-react";
-import Link from "next/link";
-
-import { getEventForms } from "./data-access";
+import { getEventFormAttributes, getEventForms } from "./data-access";
+import { EventFormForm } from "./event-form-form";
 import { FormEntry } from "./form-entry";
 
 export default async function DashboardEventFormsPage({
   params,
 }: {
-  params: { id: string };
+  params: Promise<{ id: string }>;
 }) {
-  const forms = await getEventForms(params.id);
+  const { id } = await params;
+  const forms = await getEventForms(id);
+  const attributes = await getEventFormAttributes(id);
 
   return (
     <div className="flex flex-col gap-8">
       <h1 className="text-3xl font-bold">Formularze</h1>
       <div className="mt-8 flex flex-wrap gap-8">
-        <Link
-          href="forms/create"
-          className="flex h-64 w-64 items-center justify-center gap-2 rounded-md border border-dotted border-muted p-4 text-muted-foreground"
-        >
-          <SquarePlus className="h-6 w-6" /> Stwórz formularz
-        </Link>
+        <EventFormForm eventId={id} attributes={attributes} />
         {forms.length > 0
           ? forms.map((form) => (
-              <FormEntry form={form} eventId={params.id} key={form.id} />
+              <FormEntry form={form} eventId={id} key={form.id} />
             ))
           : null}
       </div>
