+++ src/app/dashboard/events/[id]/layout.tsx
@@ -2,11 +2,9 @@ import { DashboardSidebar } from "@/components/dashboard-sidebar";
 
 export default async function DashboardEventLayout({
   children,
-  modals,
   params,
 }: {
   children: React.ReactNode;
-  modals: React.ReactNode;
   params: Promise<{ id: string }>;
 }) {
   const awaited = await params;
@@ -14,7 +12,6 @@ export default async function DashboardEventLayout({
     <div className="flex grow gap-14">
       <DashboardSidebar id={awaited.id} />
       <div className="grow">{children}</div>
-      {modals}
     </div>
   );
 }
